UI Works with different textures!!!
I'm frankly embarrassed by some of the code that I hacked out in this
commit, I did not know what I was doing there, and didn't try to figure
it out. In any case, this version is much more robust and hopefully will
be extensible.

Needs a CSS overhaul to move the CSS into the component files.